+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Added benchmark scripts.
 - Added webpack config to build js file for browser.
 - Added `clvm_tools.go(...args)` function to dispatch cli commands from javascript.
+- Added `--experiment-backend rust` option to use `clvm_rs`.
 ### Fixed
 - Fixed an issue where it did not correctly handle signed/unsigned integer from and to `Bytes`.
 - Fixed an issue where large int was not recognized correctly.